app/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.11.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.28.0
atlassian-python-api==3.39.0
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
cobble==0.1.3
dataclasses-json==0.5.13
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.4.0
faiss-cpu==1.7.4
fastapi==0.100.0
fastapi-restful==0.4.5
filelock==3.11.0
frozenlist==1.4.0
fsspec==2023.6.0
google-api-core==2.11.0
google-api-python-client==2.94.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.239
langsmith==0.0.14
lit==16.0.1
loguru==0.7.0
lxml==4.9.2
Mako==1.2.4
mammoth==1.6.0
MarkupSafe==2.1.2
marshmallow==3.20.1
mattermostdriver==7.3.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauth2client==4.1.3
oauthlib==3.2.2
openapi-schema-pydantic==1.2.4
packaging==23.1
persist-queue==0.8.1
persistQueue==0.1.6
Pillow==9.5.0
pinecone-client==2.2.2
posthog==3.0.1
protobuf==4.22.3
psutil==5.9.4
psycopg2-binary==2.9.6
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.18.0
pydantic==1.10.7
pyparsing==3.0.9
pypdf==3.12.2
pypdf2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
rank-bm25==0.2.2
regex==2023.3.23
requests==2.31.0
requests-oauthlib==1.3.1
retry==0.9.2
rocketchat-API==1.30.0
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.98
six==1.16.0
slack-sdk==3.21.3
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.19
starlette==0.27.0
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
triton==2.0.0
typing-extensions==4.5.0
typing-inspect==0.9.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.23.1
websockets==11.0.1
wrapt==1.15.0
XlsxWriter==3.1.0
yarl==1.9.2
zipp==3.16.2
openai==0.27.8